pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "masters-vlm"
version = "0.1.0"
description = "Mitigating Chain-of-Thought Overload in Small VLMs through RL Post-Training"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "strongich"}
]
dependencies = [
    "accelerate>=1.11.0",
    "datasets>=4.5.0",
    "matplotlib>=3.10.8",
    "huggingface-hub>=0.34.0,<1.0",
    "pillow>=11.3.0",
    "torch>=2.9.0",
    "trackio>=0.16.1",
    "transformers>=4.57.6",
    "deepspeed>=0.18.6",
    "trl>=0.28.0",
]

[project.optional-dependencies]
vllm = [
    "vllm",
]
dev = [
    "pytest",
    "black",
    "ruff",
]

[tool.setuptools]
packages = ["src"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.jinja"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]
